создание UserAccessManagementService и перенос туда проверки на подбор пароля и активации пользователей. Правка названия таблицы в бд (Александр говорил)
