+++ auth-service/src/main/java/com/canban/auth/controller/AuthController.java
@@ -4,6 +4,7 @@ import com.canban.api.auth.JwtRequest;
 import com.canban.api.auth.JwtResponse;
 import com.canban.auth.entity.User;
 import com.canban.auth.exceptions.InvalidAuthorizationException;
+import com.canban.auth.service.UserAccessManagementService;
 import com.canban.auth.service.UserService;
 import com.canban.auth.util.JwtTokenUtil;
 import io.swagger.v3.oas.annotations.Operation;
@@ -32,8 +33,7 @@ public class AuthController {
     private final UserService userService;
     private final JwtTokenUtil jwtTokenUtil;
     private final AuthenticationManager authenticationManager;
-
-    private List<User> users = new ArrayList<>();
+    private final UserAccessManagementService userAccessManagementService;
 
     @PostMapping("/auth")
     @Operation(
@@ -47,7 +47,7 @@ public class AuthController {
         try {
             authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(authRequest.getUsername(), authRequest.getPassword()));
         } catch (BadCredentialsException e) {
-            if (passwordGuessingProtection(authRequest.getUsername())) {
+            if (userAccessManagementService.passwordGuessingProtection(authRequest.getUsername())) {
                 log.warn("Пользователь " + authRequest.getUsername() + " пытается подобрать пароль");
             }
             throw new InvalidAuthorizationException("Incorrect username or password");
@@ -57,33 +57,4 @@ public class AuthController {
         String token = jwtTokenUtil.generateToken(userDetails);
         return ResponseEntity.ok(new JwtResponse(token));
     }
-
-    /**
-     * Проверка на попытку подбора пароля к аккаунту
-     */
-    private boolean passwordGuessingProtection(String username) {
-        User user = null;
-        boolean isFinded = false;
-        for (int i = 0; i < users.size(); i++) {
-            if (users.get(i).getUsername().equals(username)) {
-                user = users.get(i);
-                isFinded = true;
-                break;
-            }
-        }
-        if (user == null)  {
-            user = new User(username);
-        }
-        user.setMistakeNum(user.getMistakeNum() + 1);
-        if (!isFinded) {
-            users.add(user);
-        }
-        if (user.getMistakeNum() >= 3) {
-            user.setMistakeNum(0);
-            return true;
-        }
-        return false;
-    }
-
-
 }
